addressed more comments from Anastassy for PR#21
-   Reformatting code
-   Removed unused layout files
-   Placed fragment files inside the view package
-   Incorporated the changes from the main branch
