+++ app/src/main/java/com/soen490chrysalis/papilio/view/BrowseFragment.kt
@@ -1,17 +1,18 @@
-package com.soen490chrysalis.papilio
+package com.soen490chrysalis.papilio.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.soen490chrysalis.papilio.R
 
 /**
  * A simple [Fragment] subclass.
- * Use the [HomeFragment.newInstance] factory method to
+ * Use the [BrowseFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class HomeFragment : Fragment()
+class BrowseFragment : Fragment()
 {
     // TODO: Rename and change types of parameters
     private var mParam1 : String? = null
@@ -28,13 +29,12 @@ class HomeFragment : Fragment()
     }
 
     override fun onCreateView(
-        inflater : LayoutInflater,
-        container : ViewGroup?,
+        inflater : LayoutInflater, container : ViewGroup?,
         savedInstanceState : Bundle?
     ) : View?
     {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false)
+        return inflater.inflate(R.layout.fragment_browse, container, false)
     }
 
     companion object
@@ -50,12 +50,12 @@ class HomeFragment : Fragment()
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment HomeFragment.
+         * @return A new instance of fragment BrowseFragment.
          */
         // TODO: Rename and change types and number of parameters
-        fun newInstance(param1 : String?, param2 : String?) : HomeFragment
+        fun newInstance(param1 : String?, param2 : String?) : BrowseFragment
         {
-            val fragment = HomeFragment()
+            val fragment = BrowseFragment()
             val args = Bundle()
             args.putString(ARG_PARAM1, param1)
             args.putString(ARG_PARAM2, param2)
