+++ app/src/main/res/layout/account_menu_fragment.xml
@@ -121,8 +121,6 @@
 
                 </LinearLayout>
 
-
-
             </LinearLayout>
 
         </LinearLayout>
