+++ app/src/main/res/layout/activity_main.xml
@@ -12,7 +12,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_marginBottom="?actionBarSize"
-        android:name="com.soen490chrysalis.papilio.HomeFragment" />
+        android:name="com.soen490chrysalis.papilio.view.HomeFragment" />
 
     <RelativeLayout
         android:layout_width="match_parent"
