+++ app/src/main/res/layout/fragment_activities.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/activities_fragment"
-    tools:context=".ActivitiesFragment">
+    tools:context=".view.ActivitiesFragment">
 
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
