+++ app/src/main/res/layout/fragment_browse.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/browse_fragment"
-    tools:context=".BrowseFragment">
+    tools:context=".view.BrowseFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <androidx.cardview.widget.CardView
