+++ app/src/main/res/xml/network_security_config.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <network-security-config>
     <domain-config cleartextTrafficPermitted="true">
-        <domain includeSubdomains="true">192.168.1.36</domain>
+        <domain includeSubdomains="true">184.160.92.157</domain>
     </domain-config>
 </network-security-config>
